add clean architecture

+++ app/src/androidTest/java/com/example/fizzbuzz/presentation/second_screen/ResultViewModelTest.kt
@@ -1,6 +1,6 @@
 package com.example.fizzbuzz.presentation.second_screen
 
-import com.example.fizzbuzz.models.GameModel
+import com.example.fizzbuzz.domain.model.GameModel
 import com.google.common.truth.Truth
 import org.junit.Test
 
@@ -10,20 +10,20 @@ class ResultViewModelTest{
 
     @Test
     fun validListSize(){
-        val gameModel=GameModel(3,6,10,"A","B")
+        val gameModel= GameModel(3,6,10,"A","B")
         val list=viewModel.showData(gameModel)
-        Truth.assertThat(list.size==gameModel.limit).isTrue()
+        Truth.assertThat(list.size.toLong()==gameModel.limit).isTrue()
     }
     @Test
     fun validFirstString(){
-        val gameModel=GameModel(3,6,10,"A","B")
+        val gameModel= GameModel(3,6,10,"A","B")
         val list=viewModel.showData(gameModel)
         Truth.assertThat(list[gameModel.firstInt-1] == gameModel.str1).isTrue()
     }
 
     @Test
     fun validSecondString(){
-        val gameModel=GameModel(3,6,10,"A","B")
+        val gameModel= GameModel(3,6,10,"A","B")
         val list=viewModel.showData(gameModel)
         Truth.assertThat(list[gameModel.secondInt-1] ==gameModel.str2).isTrue()
     }
